+++ source/src/python/PyRosetta/src/pyrosetta/distributed/cluster/tools.py
@@ -56,7 +56,7 @@ from pyrosetta.distributed.cluster.core import PyRosettaCluster
 from pyrosetta.distributed.packed_pose.core import PackedPose
 from pyrosetta.rosetta.basic import was_init_called
 from pyrosetta.rosetta.core.pose import Pose
-from pyrosetta.utility.initialization import PyRosettaIsNotInitializedError
+from pyrosetta.utility.exceptions import PyRosettaIsNotInitializedError
 from typing import (
     Any,
     Callable,
